sistema_estacao/requirements.txt
Django>=5.0,<6.0
pyserial>=3.5
